+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Presets were removed in favour of the bound brushes that were added.
 - The undo/redo stores are no longer automatically cleared after 3 minutes. This caused unexpected behaviour and is never expected behaviour.
 - Templates have been removed and are no longer supported. They were deprecated for a long time since schematics were more efficient.
+- Cloning no longer operates in different shapes, but always clones the entire selection.
 
 ### Fixed
 - Fixed cuboid and cylinder shapes not to be async/sync when they should be.